+++ guillotina/configuration/tests/notyet.py
@@ -1,1 +1,0 @@
-# We don't get imported automagically