+++ guillotina/configuration/tests/samplepackage/NamedForClass.py
@@ -1,3 +1,0 @@
-# Test the "repeat" feature of guillotina.configuration.name.resolve.
-class NamedForClass(object):
-    pass